Story + lvl1 valmis

+++ Bullet Hell/Assets/Scripts/Score1.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 
 public class Score : MonoBehaviour
 {
-    public static int score = 0;
+    public static float score = 0;
     public TMP_Text scoreText;
 
     // Update is called once per frame
